Handle validation in generalized background task
