+++ androidvalidatedforms/src/main/java/com/azavea/androidvalidatedforms/FormActivityBase.java
@@ -1,6 +1,8 @@
 package com.azavea.androidvalidatedforms;
 
 /**
+ * Expected interface for FormWithAppCompatActivity and FormActivity.
+ *
  * Created by kathrynkillebrew on 12/29/15.
  */
 public interface FormActivityBase {
